+++ lib/go-atscfg/volumedotconfig.go
@@ -24,6 +24,7 @@ import (
 )
 
 const ContentTypeVolumeDotConfig = ContentTypeTextASCII
+const LineCommentVolumeDotConfig = LineCommentHash
 
 // MakeVolumeDotConfig creates volume.config for a given ATS Profile.
 // The paramData is the map of parameter names to values, for all parameters assigned to the given profile, with the config_file "storage.config".