+++ traffic_ops/traffic_ops_golang/api/shared_handlers.go
@@ -114,8 +114,8 @@ func decodeAndValidateRequestBody(r *http.Request, v Validator) error {
 	return v.Validate()
 }
 
-func hasDeleteKeyOption(obj interface{}, params map[string]string) (bool, error, error, int) {
-	optionsDeleter, ok := obj.(HasDeleteKeyOptions)
+func checkIfOptionsDeleter(obj interface{}, params map[string]string) (bool, error, error, int) {
+	optionsDeleter, ok := obj.(OptionsDeleter)
 	if !ok {
 		return false, nil, nil, http.StatusOK
 	}
@@ -310,14 +310,38 @@ func DeleteHandler(deleter Deleter) http.HandlerFunc {
 		obj := reflect.New(objectType).Interface().(Deleter)
 		obj.SetInfo(inf)
 
-		deleteKeyOptionExists, userErr, sysErr, errCode := hasDeleteKeyOption(obj, inf.Params)
+		isOptionsDeleter, userErr, sysErr, errCode := checkIfOptionsDeleter(obj, inf.Params)
 		if userErr != nil || sysErr != nil {
 			HandleErr(w, r, inf.Tx.Tx, errCode, userErr, sysErr)
 			return
 		}
-		if !deleteKeyOptionExists {
+		var (
+			keys = make(map[string]interface{})
+			err  error
+		)
+		if isOptionsDeleter {
+			for key, info := range obj.(OptionsDeleter).DeleteKeyOptions() {
+				paramKey := inf.Params[key]
+				if paramKey == "" {
+					continue
+				}
+				switch reflect.ValueOf(info.Checker) {
+				case reflect.ValueOf(IsInt):
+					if keys[key], err = GetIntKey(paramKey); err != nil {
+						HandleErr(w, r, inf.Tx.Tx, http.StatusBadRequest, errors.New("failed to parse key: "+key), nil)
+						return
+					}
+				case reflect.ValueOf(IsBool):
+					if keys[key], err = strconv.ParseBool(paramKey); err != nil {
+						HandleErr(w, r, inf.Tx.Tx, http.StatusBadRequest, errors.New("failed to parse key: "+key), nil)
+						return
+					}
+				default:
+					keys[key] = paramKey
+				}
+			}
+		} else {
 			keyFields := obj.GetKeyFieldsInfo() // expecting a slice of the key fields info which is a struct with the field name and a function to convert a string into a interface{} of the right type. in most that will be [{Field:"id",Func: func(s string)(interface{},error){return strconv.Atoi(s)}}]
-			keys := make(map[string]interface{})
 			for _, kf := range keyFields {
 				paramKey := inf.Params[kf.Field]
 				if paramKey == "" {
@@ -332,8 +356,8 @@ func DeleteHandler(deleter Deleter) http.HandlerFunc {
 				}
 				keys[kf.Field] = paramValue
 			}
-			obj.SetKeys(keys) // if the type assertion of a key fails it will be should be set to the zero value of the type and the delete should fail (this means the code is not written properly no changes of user input should cause this.)
 		}
+		obj.SetKeys(keys) // if the type assertion of a key fails it will be should be set to the zero value of the type and the delete should fail (this means the code is not written properly no changes of user input should cause this.)
 
 		if t, ok := obj.(Tenantable); ok {
 			authorized, err := t.IsTenantAuthorized(inf.User)
@@ -347,7 +371,7 @@ func DeleteHandler(deleter Deleter) http.HandlerFunc {
 			}
 		}
 
-		if deleteKeyOptionExists {
+		if isOptionsDeleter {
 			obj := reflect.New(objectType).Interface().(OptionsDeleter)
 			obj.SetInfo(inf)
 			userErr, sysErr, errCode = obj.OptionsDelete()
@@ -392,14 +416,38 @@ func DeprecatedDeleteHandler(deleter Deleter, alternative *string) http.HandlerF
 		obj := reflect.New(objectType).Interface().(Deleter)
 		obj.SetInfo(inf)
 
-		deleteKeyOptionExists, userErr, sysErr, errCode := hasDeleteKeyOption(obj, inf.Params)
+		isOptionsDeleter, userErr, sysErr, errCode := checkIfOptionsDeleter(obj, inf.Params)
 		if userErr != nil || sysErr != nil {
 			HandleDeprecatedErr(w, r, inf.Tx.Tx, errCode, userErr, sysErr, alternative)
 			return
 		}
-		if !deleteKeyOptionExists {
+		var (
+			keys = make(map[string]interface{})
+			err  error
+		)
+		if isOptionsDeleter {
+			for key, info := range obj.(OptionsDeleter).DeleteKeyOptions() {
+				paramKey := inf.Params[key]
+				if paramKey == "" {
+					continue
+				}
+				switch reflect.ValueOf(info.Checker) {
+				case reflect.ValueOf(IsInt):
+					if keys[key], err = GetIntKey(paramKey); err != nil {
+						HandleDeprecatedErr(w, r, inf.Tx.Tx, http.StatusBadRequest, errors.New("missing key: "+key), nil, alternative)
+						return
+					}
+				case reflect.ValueOf(IsBool):
+					if keys[key], err = strconv.ParseBool(paramKey); err != nil {
+						HandleDeprecatedErr(w, r, inf.Tx.Tx, http.StatusBadRequest, errors.New("failed to parse key: "+key), nil, alternative)
+						return
+					}
+				default:
+					keys[key] = paramKey
+				}
+			}
+		} else {
 			keyFields := obj.GetKeyFieldsInfo() // expecting a slice of the key fields info which is a struct with the field name and a function to convert a string into a interface{} of the right type. in most that will be [{Field:"id",Func: func(s string)(interface{},error){return strconv.Atoi(s)}}]
-			keys := make(map[string]interface{})
 			for _, kf := range keyFields {
 				paramKey := inf.Params[kf.Field]
 				if paramKey == "" {
@@ -414,8 +462,8 @@ func DeprecatedDeleteHandler(deleter Deleter, alternative *string) http.HandlerF
 				}
 				keys[kf.Field] = paramValue
 			}
-			obj.SetKeys(keys) // if the type assertion of a key fails it will be should be set to the zero value of the type and the delete should fail (this means the code is not written properly no changes of user input should cause this.)
 		}
+		obj.SetKeys(keys) // if the type assertion of a key fails it will be should be set to the zero value of the type and the delete should fail (this means the code is not written properly no changes of user input should cause this.)
 
 		if t, ok := obj.(Tenantable); ok {
 			authorized, err := t.IsTenantAuthorized(inf.User)
@@ -429,7 +477,7 @@ func DeprecatedDeleteHandler(deleter Deleter, alternative *string) http.HandlerF
 			}
 		}
 
-		if deleteKeyOptionExists {
+		if isOptionsDeleter {
 			obj := reflect.New(objectType).Interface().(OptionsDeleter)
 			obj.SetInfo(inf)
 			userErr, sysErr, errCode = obj.OptionsDelete()